routing: forget info for contradictions
If we correct the success/fail amount for capacity, this means that a
new smaller/larger channel is now active. Only in the case where we
receive a contradiction, for example in the case where both amounts are
set to the capacity, we'll forget all previous experience about the
channel, as it's likely that our previous observation is outdated.