routing: add fee level setter/getter to mc state
We track the fee level per node as a map from pubkey to fee level. The
fee level's value is dependent on the last observation date and is
designed to decay with time. The reason for this is to not let the fee
level grow unbounded over time and to lower the fee level in order to
give cheaper channels another chance.