routing: add fee level delta computation to result interpretation
Some routing nodes or targets are hard to reach because they lack
inbound liquidity in low fee channels. This can be tackled by assigning
a fee level to each node giving a threshold to skip the node's channels
below that fee level. The fee level is determined iteratively and bumped
up on each failure if a node is involved in a route and the bump size is
determined by the code added here.

Up to now the failure of a hop only gives a local penalty to the hop's
probability and to some extent to the "node probability" (it quantifies
how good that node is in forwarding outwards). However, this is not
effective enough to remove the most illiquid routes, leading to many
trial and error attempts.

To address this, when encountering a temporary channel failure (i.e.
liquidity problems), we also make an inference on the downstream nodes
by bumping up their fee level. The fee level bump function is designed
such that the node that we failed to reach has its fee level increased
the most, declining with some decay length towards the ends of the route
for other nodes. This mechanism ensures that we will not focus on the
low fee channels of the target node further. The usage of the fee level
in pathfinding is subject of a later commit.

Optionally we also bump the fee level in the successful part of the
route, since those hop's fee rates misguided the pathfinding algorithm
to choose a route via nodes that couldn't forward the payment.