routing: use fee level in pathfinding
In this commit we adapt pathfinding to focus less on the channels that
are below the fee level. This is done by changing the edge weight of
channels that charge an effective fee (base fee and fee rate) that is
below the equivalent fee level.

The real fee for such channels is changed such that its virtual fee is
the one mirrored around the fee level. This way we still consider those
channels, but at higher fee rates. By the mirroring operation, the
lowest fee channels will be bumped up the most, while channels slightly
below the fee level stay roughly in the same range above. This discourages the
usage of low fee channels while still considering them as an alternative
in routing, blended into the fee range of higher-fee channels.